Change Type() signature to return interface and error (#78)
This PR changes TypeMatcher's Type() signature to return interface and error.

This change is needed to address an issue 7714 where the offline node is receiving messages for queue work (work ready, work complete), but it is not configured to handle those message types. This error is innocuous, but could be alarming to customers.