feat: option to initialize storage servers without chunker
The goal with this change is to allow storage servers
to be initialized without needing a database to be running
first.

By allowing the storage server to start without the chunker,
it can still use Put() without chunking. When WriteChunked()
or ReadChunked() are needed by the storage server, it first
checks that the chunker is not nil. If it is, it will throw
an error.